Выполнил задание по модулям. Разбил код на модули
