+++ js/create-author.js
@@ -1,0 +1,8 @@
+import {getRandomNumber} from './util.js';
+const createAuthor = () => {
+  return {
+    avatar: `img/avatars/user0${getRandomNumber(1, 8)}.png`,
+  };
+};
+
+export {createAuthor};
