+++ js/get-random-element.js
@@ -1,0 +1,6 @@
+//import _ from ('lodash');
+const getRandomElement = (elementsArray) => {
+  return _.sample([elementsArray]);
+};
+
+export {getRandomElement};
